% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxPersistencePara2.R
\name{maxPersistencePara2}
\alias{maxPersistencePara2}
\title{Fully parallel Version of Maximal Persistence Method}
\usage{
maxPersistencePara2(
  FUN,
  parameters,
  X,
  lim,
  by,
  Grid,
  maxdimension = length(lim)/2 - 1,
  sublevel = TRUE,
  library = "GUDHI",
  B = 30,
  alpha = 0.05,
  bandFUN = "bootstrapBand",
  distance = "bottleneck",
  dimension = min(1, maxdimension),
  p = 1,
  printProgress = FALSE,
  corePermit = parallel::detectCores(),
  weight = NULL
)
}
\arguments{
\item{FUN}{the distance function adopted for filtration, such as \code{dtm}, \code{kde}, etc.}

\item{parameters}{a numeric vector for tuning. The optimal choice among them will be chosen.}

\item{X}{the dataset, a matrix.}

\item{lim}{the range of each dimension of the grid.}

\item{by}{the resolution of the grid.}

\item{Grid}{the grid expanded.}

\item{maxdimension}{the maximum dimension to compute PH.}

\item{sublevel}{a logical value, whether the sublevel set PH should be computed.}

\item{library}{a string specifying which library should be used to compute.}

\item{B}{the number of bootstraps.}

\item{alpha}{confident level}

\item{bandFUN}{the function to be used in the computation, either \code{"bootstrapDiagram"} or \code{"bootstrapBand"}.}

\item{distance}{the same as \code{maxPersistence}.}

\item{dimension}{the same as \code{maxPersistence}.}

\item{p}{the same as \code{maxPersistence}.}

\item{printProgress}{the same as \code{maxPersistence}.}

\item{corePermit}{the number of CPU cores permitted for tuning.}

\item{weight}{the same as \code{maxPersistence}.}
}
\description{
The same function detecting maximal persistence as that of \code{TDA} package. In the function multiple computation kernel would be opened and parallel computation will be conducted using all of them. Further, bootstrap process involves CPU cores at rest.
}
\details{
This function is the parallel version of the TDA::maxPersistence
for the parallel computation of the parameter.
This function differs the \code{masPersistencePara} in inner tuning mechanism:
the function allows the core at rest to further join bootstrapping process.
}
