% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxPersistenceParApprox.R
\name{maxPersistenceParApprox}
\alias{maxPersistenceParApprox}
\title{Parallel Version of Maximal Persistence Method (Random Sampling Estimation)}
\usage{
maxPersistenceParApprox(
  FUN,
  parameters,
  X,
  lim,
  by,
  approxSize = nrow(X),
  maxdimension = length(lim)/2 - 1,
  sublevel = TRUE,
  library = "GUDHI",
  B = 30,
  alpha = 0.05,
  bandFUN = "sampleBand",
  distance = "bottleneck",
  dimension = min(1, maxdimension),
  p = 1,
  printProgress = FALSE,
  weight = NULL,
  corePermit = parallel::detectCores(),
  os = "windows"
)
}
\arguments{
\item{FUN}{the distance function adopted for filtration, such as \code{dtm}, \code{kde}, etc.}

\item{parameters}{a numeric vector for tuning. The optimal choice among them will be chosen.}

\item{X}{the dataset, a matrix.}

\item{lim}{the range of each dimension of the grid.}

\item{by}{the resolution of the grid.}

\item{maxdimension}{the maximum dimension to compute PH.}

\item{sublevel}{a logical value, whether the sublevel set PH should be computed.}

\item{library}{a string specifying which library should be used to compute.}

\item{B}{the number of bootstraps.}

\item{alpha}{confident level}

\item{bandFUN}{the function to be used in the computation, either \code{"bootstrapDiagram"}, \code{"bootstrapBand"} or \code{"sampleDiagram"}.}

\item{distance}{the same as \code{maxPersistence}.}

\item{dimension}{the same as \code{maxPersistence}.}

\item{p}{the same as \code{maxPersistence}.}

\item{printProgress}{the same as \code{maxPersistence}.}

\item{weight}{the same as \code{maxPersistence}.}

\item{corePermit}{the number of CPU cores permitted for tuning.}

\item{os}{the OS.}
}
\description{
The same function detecting maximal persistence as that of \code{TDA} package. In the function multiple computation kernel would be opened and parallel computation will be conducted using all of them.
}
\details{
This function is similar to maxPersistencePara, yet only randomly extract a sample from the dataset
to estimate the confidence band. Parallel computation is also adopted.
}
