% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBand.R
\name{sampleBand}
\alias{sampleBand}
\title{Sampling Confidence Band}
\usage{
sampleBand(
  X,
  FUN,
  Grid,
  sampleSize,
  B = 30,
  alpha = 0.05,
  parallel = FALSE,
  printProgress = FALSE,
  weight = NULL,
  ...
)
}
\arguments{
\item{X}{the dataset, a matrix.}

\item{FUN}{the distance function adopted for filtration, such as \code{dtm}, \code{kde}, etc.}

\item{Grid}{the meshgrid used for computation.}

\item{sampleSize}{the sample size of the X during sampling step.}

\item{B}{the number of bootstraps.}

\item{alpha}{confident level}

\item{printProgress}{the same as \code{maxPersistence}.}

\item{weight}{the same as  \code{maxPersistence}.}
}
\description{
This function is the confidence estimator similar to the \code{TDA::bootstrapBand},
usually for big dataset where full size analysis is computationally intractable.
}
