% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDiagram.R
\name{sampleDiagram}
\alias{sampleDiagram}
\title{Sampling Confidence Set for a Persistence Diagram, using the Bottleneck Distance (or the Wasserstein distance).}
\usage{
sampleDiagram(
  X,
  FUN,
  lim,
  by,
  sampleSize = nrow(X),
  maxdimension = length(lim)/2 - 1,
  sublevel = TRUE,
  library = "GUDHI",
  B = 30,
  alpha = 0.05,
  distance = "bottleneck",
  dimension = min(1, maxdimension),
  p = 1,
  parallel = FALSE,
  printProgress = FALSE,
  weight = NULL,
  ...
)
}
\arguments{
\item{X}{the dataset, a matrix.}

\item{FUN}{the distance function adopted for filtration, such as \code{dtm}, \code{kde}, etc.}

\item{lim}{the range of each dimension of the grid.}

\item{by}{the resolution of the grid.}

\item{sampleSize}{the sample size of the X during sampling step.}

\item{maxdimension}{the maximum dimension to compute PH.}

\item{sublevel}{a logical value, whether the sublevel set PH should be computed.}

\item{library}{a string specifying which library should be used to compute.}

\item{B}{the number of bootstraps.}

\item{alpha}{confident level}

\item{distance}{the same as \code{maxPersistence}.}

\item{dimension}{the same as \code{maxPersistence}.}

\item{p}{the same as \code{maxPersistence}.}

\item{printProgress}{the same as \code{maxPersistence}.}

\item{weight}{the same as  \code{maxPersistence}.}

\item{bandFUN}{the function to be used in the computation, either \code{"bootstrapDiagram"}, \code{"bootstrapBand"} or \code{"sampleDiagram"}.}
}
\description{
This function is the confidence estimator similar to the \code{TDA::bootstrapDiagram},
usually for big dataset where full size analysis is computationally intractable.
}
